refactor: code under src/ + bump go version 1.21
