+++ .github/workflows/release.yaml
@@ -43,7 +43,7 @@ jobs:
         name: Set up Go
         uses: actions/setup-go@v2
         with:
-          go-version: 1.19
+          go-version: 1.21
       
       # - name: dockerhub-login
       #   uses: docker/login-action@v1
