+++ src/internal/crypto/jwt/vault/transit/es256.go
