+++ src/internal/crypto/jwt/vault/transit/es512.go
