+++ src/internal/crypto/jwt/vault/transit/key.go
