+++ src/internal/crypto/jwt/vault/transit/main.go
