+++ src/internal/crypto/jwt/vault/transit/rs256.go
