+++ src/internal/crypto/jwt/vault/transit/rs384.go
