+++ src/internal/crypto/jwt/vault/transit/rs512.go
